update the build process
I'm adding DEBUG_OUTPUT_PATH and RELEASE_OUTPUT_PATH to the build
process as well as a default value for SUBNAUTICA_PATH.

This to allow for further local control on where the build result ends
up and where you store this git repository.

Additionally, during a Release build the result will also be copied to
the Debug path so they can be tested in-game.