update the build process
I'm adding DEBUG_OUTPUT_PATH and RELEASE_OUTPUT_PATH to the build
process as well as SUBNAUTICA_PATH - removing the hard coded paths
that prevented me from building the project on my computer.

- When building with the Debug configuration, the results will be
placed directly in the DEBUG_OUTPUT_PATH.
- During a Release build the results will be written to the
RELEASE_OUTPUT_PATH but they will also be copied to the Debug path
so they can be tested in-game.

I've added a simplistic description of the process in README.md